Move wellknown prop to its package
