+++ src/main/kotlin/de/bund/digitalservice/useid/wellknown/WellKnownProperties.kt
@@ -1,4 +1,4 @@
-package de.bund.digitalservice.useid.config
+package de.bund.digitalservice.useid.wellknown
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.stereotype.Component
